--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -1,12 +1,18 @@
 package hw10programoptimization
 
 import (
-	"encoding/json"
+	"bufio"
+	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
-	"regexp"
 	"strings"
+
+	"github.com/mailru/easyjson"
+)
+
+var (
+	ErrInvalidUser  = errors.New("invalid user")
+	ErrInvalidEmail = errors.New("invalid email")
 )
 
 type User struct {
@@ -22,46 +28,31 @@ type User struct {
 type DomainStat map[string]int
 
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
-	u, err := getUsers(r)
-	if err != nil {
-		return nil, fmt.Errorf("get users error: %w", err)
-	}
-	return countDomains(u, domain)
-}
-
-type users [100_000]User
-
-func getUsers(r io.Reader) (result users, err error) {
-	content, err := ioutil.ReadAll(r)
-	if err != nil {
-		return
-	}
-
-	lines := strings.Split(string(content), "\n")
-	for i, line := range lines {
-		var user User
-		if err = json.Unmarshal([]byte(line), &user); err != nil {
-			return
-		}
-		result[i] = user
-	}
-	return
-}
-
-func countDomains(u users, domain string) (DomainStat, error) {
 	result := make(DomainStat)
+	suffix := "." + strings.ToLower(domain)
 
-	for _, user := range u {
-		matched, err := regexp.Match("\\."+domain, []byte(user.Email))
-		if err != nil {
-			return nil, err
+	scanner := bufio.NewScanner(r)
+	for scanner.Scan() {
+		var user User
+		if err := easyjson.Unmarshal(scanner.Bytes(), &user); err != nil {
+			return nil, fmt.Errorf("%w: %v", ErrInvalidUser, err)
 		}
 
-		if matched {
-			num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
-			num++
-			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
+		email := strings.ToLower(user.Email)
+		if matched := strings.HasSuffix(email, suffix); matched {
+			d, err := extractDomain(email)
+			if err != nil {
+				return nil, fmt.Errorf("extract domain: %w", err)
+			}
+			result[d]++
 		}
 	}
 	return result, nil
+}
+
+func extractDomain(email string) (string, error) {
+	if parts := strings.SplitN(email, "@", 2); len(parts) == 2 {
+		return parts[1], nil
+	}
+	return "", ErrInvalidEmail
 }