--- a/hw10_program_optimization/stats_test.go
+++ b/hw10_program_optimization/stats_test.go
@@ -1,4 +1,4 @@
-// +build !bench
+//go:build !bench
 
 package hw10programoptimization
 
@@ -35,5 +35,49 @@ func TestGetDomainStat(t *testing.T) {
 		result, err := GetDomainStat(bytes.NewBufferString(data), "unknown")
 		require.NoError(t, err)
 		require.Equal(t, DomainStat{}, result)
+	})
+}
+
+func TestGetDomainStat_InvalidData(t *testing.T) {
+	t.Run("invalid user", func(t *testing.T) {
+		_, err := GetDomainStat(bytes.NewBufferString("invalid data"), "com")
+		require.ErrorIs(t, err, ErrInvalidUser)
+	})
+	t.Run("invalid email", func(t *testing.T) {
+		data := `{"Id":1,"Name":"n","Username":"un","Email":"unexample.com","Phone":"0","Password":"p","Address":"a"}`
+		_, err := GetDomainStat(bytes.NewBufferString(data), "com")
+		require.ErrorIs(t, err, ErrInvalidEmail)
+	})
+}
+
+func BenchmarkGetDomain(b *testing.B) {
+	data := `{"Id":1,"Name":"Howard Mendoza","Username":"0Oliver","Email":"aliquid_qui_ea@Browsedrive.gov","Phone":"6-866-899-36-79","Password":"InAQJvsq","Address":"Blackbird Place 25"}
+{"Id":2,"Name":"Jesse Vasquez","Username":"qRichardson","Email":"mLynch@broWsecat.com","Phone":"9-373-949-64-00","Password":"SiZLeNSGn","Address":"Fulton Hill 80"}
+{"Id":3,"Name":"Clarence Olson","Username":"RachelAdams","Email":"RoseSmith@Browsecat.com","Phone":"988-48-97","Password":"71kuz3gA5w","Address":"Monterey Park 39"}
+{"Id":4,"Name":"Gregory Reid","Username":"tButler","Email":"5Moore@Teklist.net","Phone":"520-04-16","Password":"r639qLNu","Address":"Sunfield Park 20"}
+{"Id":5,"Name":"Janice Rose","Username":"KeithHart","Email":"nulla@Linktype.com","Phone":"146-91-01","Password":"acSBF5","Address":"Russell Trail 61"}`
+
+	b.Run("find 'com'", func(b *testing.B) {
+		for i := 0; i < b.N; i++ {
+			if _, err := GetDomainStat(bytes.NewBufferString(data), "com"); err != nil {
+				b.Fatal(err)
+			}
+		}
+	})
+
+	b.Run("find 'gov'", func(b *testing.B) {
+		for i := 0; i < b.N; i++ {
+			if _, err := GetDomainStat(bytes.NewBufferString(data), "gov"); err != nil {
+				b.Fatal(err)
+			}
+		}
+	})
+
+	b.Run("find 'unknown'", func(b *testing.B) {
+		for i := 0; i < b.N; i++ {
+			if _, err := GetDomainStat(bytes.NewBufferString(data), "unknown"); err != nil {
+				b.Fatal(err)
+			}
+		}
 	})
 }