## Домашнее задание №10 «Оптимизация программы»

### Чек-лист студента ([Что это?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0))
- [x] Я перечитал задание после решения. [Зачем перечитывать задание?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0#user-content-%D0%97%D0%B0%D1%87%D0%B5%D0%BC-%D0%BF%D0%B5%D1%80%D0%B5%D1%87%D0%B8%D1%82%D1%8B%D0%B2%D0%B0%D1%82%D1%8C-%D0%B7%D0%B0%D0%B4%D0%B0%D0%BD%D0%B8%D0%B5)
- [x] Я запустил `go mod tidy`.
- [x] Я удалил `.sync` файл. [Зачем его удалять?](https://github.com/OtusGolang/home_work/wiki/%5B%D0%A1%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0%D0%BC%5D-%D0%9F%D1%80%D0%BE%D1%86%D0%B5%D1%81%D1%81-%D1%81%D0%B4%D0%B0%D1%87%D0%B8-%D0%94%D0%97#user-content-%D0%92%D0%B0%D1%80%D0%B8%D0%B0%D0%BD%D1%82-2)
- [x] Я написал бенчмарк и приложил результаты benchstat.

### Критерии оценки
- [x] Пайплайн зелёный и нет попытки «обмануть» систему - 4 балла
- [x] Добавлены юнит-тесты - до 3 баллов
- [x] Понятность и чистота кода - до 3 баллов

#### Зачёт от 7 баллов


### benchstat

## old
```
name                        time/op
GetDomain/find_'com'-8      156ms ± 1%
GetDomain/find_'gov'-8      155ms ± 1%
GetDomain/find_'unknown'-8  209ms ± 2%

name                        alloc/op
GetDomain/find_'com'-8      136MB ± 0%
GetDomain/find_'gov'-8      136MB ± 0%
GetDomain/find_'unknown'-8  204MB ± 0%

name                        allocs/op
GetDomain/find_'com'-8      1.70M ± 0%
GetDomain/find_'gov'-8      1.70M ± 0%
GetDomain/find_'unknown'-8  1.90M ± 0%
```

## new
```
name                        time/op
GetDomain/find_'com'-8      5.96µs ± 6%
GetDomain/find_'gov'-8      5.48µs ± 3%
GetDomain/find_'unknown'-8  5.37µs ± 8%

name                        alloc/op
GetDomain/find_'com'-8      6.01kB ± 0%
GetDomain/find_'gov'-8      5.94kB ± 0%
GetDomain/find_'unknown'-8  5.70kB ± 0%

name                        allocs/op
GetDomain/find_'com'-8        43.0 ± 0%
GetDomain/find_'gov'-8        41.0 ± 0%
GetDomain/find_'unknown'-8    39.0 ± 0%
```

## delta
```
name                        old time/op    new time/op    delta
GetDomain/find_'com'-8         156ms ± 1%       0ms ± 6%  -100.00%  (p=0.008 n=5+5)
GetDomain/find_'gov'-8         155ms ± 1%       0ms ± 3%  -100.00%  (p=0.008 n=5+5)
GetDomain/find_'unknown'-8     209ms ± 2%       0ms ± 8%  -100.00%  (p=0.008 n=5+5)

name                        old alloc/op   new alloc/op   delta
GetDomain/find_'com'-8         136MB ± 0%       0MB ± 0%  -100.00%  (p=0.008 n=5+5)
GetDomain/find_'gov'-8         136MB ± 0%       0MB ± 0%  -100.00%  (p=0.008 n=5+5)
GetDomain/find_'unknown'-8     204MB ± 0%       0MB ± 0%  -100.00%  (p=0.008 n=5+5)

name                        old allocs/op  new allocs/op  delta
GetDomain/find_'com'-8         1.70M ± 0%     0.00M ± 0%  -100.00%  (p=0.008 n=5+5)
GetDomain/find_'gov'-8         1.70M ± 0%     0.00M ± 0%  -100.00%  (p=0.008 n=5+5)
GetDomain/find_'unknown'-8     1.90M ± 0%     0.00M ± 0%  -100.00%  (p=0.008 n=5+5)
```